requirements.txt
streamlit>=1.32.0
langchain>=0.1.13
langchain-openai>=0.1.0
langchain-community>=0.0.29
SQLAlchemy>=2.0.29
pandas>=2.2.0
python-dotenv>=1.0.1
openpyxl>=3.1.5
